fix(lint): resolve all golangci-lint issues
- gofmt: reformat configmaps.go, deployments.go, hpa.go
- prealloc: preallocate slices in networkpolicies.go, validation.go
- ginkgolinter: use HaveLen/BeEmpty/HavePrefix instead of len() assertions
- unconvert: remove redundant int32 casts in deployments_test.go
- unparam: drop always-nil error return from mcpgateway_test helpers
- unused: remove dead componentsNeedingNSRole variable
- goconst: extract testAPIServerCIDR constant, nolint test-only "10m"
- gocyclo: extract deployAPIFrontendExtras to reduce deployWorkloads complexity

+++ internal/resources/deployments_test.go
@@ -957,9 +957,9 @@ var _ = Describe("APIFrontendDeployment", func() {
 		for _, p := range container.Ports {
 			portMap[p.Name] = p.ContainerPort
 		}
-		Expect(portMap).To(HaveKeyWithValue("https", int32(PortHTTPS)))
-		Expect(portMap).To(HaveKeyWithValue("health", int32(PortHealthProbe)))
-		Expect(portMap).To(HaveKeyWithValue("metrics", int32(PortMetrics)))
+		Expect(portMap).To(HaveKeyWithValue("https", PortHTTPS))
+		Expect(portMap).To(HaveKeyWithValue("health", PortHealthProbe))
+		Expect(portMap).To(HaveKeyWithValue("metrics", PortMetrics))
 	})
 
 	It("mounts config, tls-server, tls-ca, and tmp volumes", func() {
